Extracts mock functions to specific file
